Run token verification only on master (#1022)
